11351 - made new component for the recipient search results; moved the api call to searchHelper.js
