+++ src/js/components/search/SearchSidebar.jsx
@@ -190,7 +190,7 @@ const SearchSidebar = ({
 
     releasedFilters.options.forEach((filter) => {
     // Collapse all by default, unless the filter has a selection made
-        if (filter.title === 'Time Period' || filter.title === 'Recipient Type' || filter.title === 'Recipient') {
+        if (filter.title === 'Time Period') {
             // time period is always expanded
             expanded.push(true);
         }
